extracts usql client
